restruture notes
