+++ src/core/Notes/Handlers/Get.cs
