+++ src/core/Notes/Handlers/NoteCreatedHandler.cs
