--- src/core/Notes/NoteEnriched.cs
@@ -1,17 +1,0 @@
-using System;
-using core.Adapters.Stocks;
-using core.Shared;
-
-namespace core.Notes
-{
-    public class NoteEnriched : AggregateEvent
-    {
-        public NoteEnriched(Guid id, Guid aggregateId, DateTimeOffset when, StockAdvancedStats stats)
-            : base(id, aggregateId, when)
-        {
-            this.Stats = stats;
-        }
-
-        public StockAdvancedStats Stats { get; }
-    }
-}
