--- src/core/Notes/NoteEnrichedWithPrice.cs
@@ -1,19 +1,0 @@
-using System;
-using core.Adapters.Stocks;
-using core.Shared;
-
-namespace core.Notes
-{
-    public class NoteEnrichedWithPrice : AggregateEvent
-    {
-        public NoteEnrichedWithPrice(Guid id, Guid aggregateId, DateTimeOffset when, TickerPrice price, StockAdvancedStats stats)
-            : base(id, aggregateId, when)
-        {
-            this.Price = price;
-            this.Stats = stats;
-        }
-
-        public TickerPrice Price { get; }
-        public StockAdvancedStats Stats { get; }
-    }
-}
