--- src/core/Notes/NoteReminderCleared.cs
@@ -1,12 +1,0 @@
-using System;
-using core.Shared;
-
-namespace core.Notes
-{
-    public class NoteReminderCleared : AggregateEvent
-    {
-        public NoteReminderCleared(Guid id, Guid aggregateId, DateTimeOffset when) : base(id, aggregateId, when)
-        {
-        }
-    }
-}
