--- src/core/Notes/NoteReminderSet.cs
@@ -1,12 +1,0 @@
-using System;
-using core.Shared;
-
-namespace core.Notes
-{
-    public class NoteReminderSet : AggregateEvent
-    {
-        public NoteReminderSet(Guid id, Guid aggregateId, DateTimeOffset when) : base(id, aggregateId, when)
-        {
-        }
-    }
-}
